add AdminController, fix PersonController
