+++ src/main/java/com/project/green/controller/PersonController.java
@@ -1,12 +1,12 @@
 package com.project.green.controller;
 
 import com.project.green.dto.PersonDto;
+import com.project.green.security.CustomUserDetails;
 import com.project.green.service.PersonService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("api/person")
@@ -16,19 +16,10 @@ public class PersonController {
     @Autowired
     private PersonService personService;
 
-    @PostMapping("/create")
-    public void create(@RequestBody PersonDto personDto) {
-        personService.save(personDto);
-    }
 
-    @GetMapping("/all")
-    public List<PersonDto> findAll() {
-        return personService.getAll();
-    }
-
-    @GetMapping("/{id}")
-    public PersonDto findById(@PathVariable("id") int id) {
-        return personService.getById(id);
+    @GetMapping("")
+    public PersonDto getPerson(@AuthenticationPrincipal CustomUserDetails userDetails) {
+        return personService.getByEmail(userDetails.getUsername());
     }
 
     @PutMapping("/update")
@@ -36,8 +27,4 @@ public class PersonController {
         return personService.update(personDto);
     }
 
-    @DeleteMapping("/delete")
-    public void delete(@RequestBody PersonDto personDto) {
-        personService.deleteById(personDto.getId());
-    }
 }
